src/ol_openedx_course_outline_api/pyproject.toml
[project]
name = "ol-openedx-course-outline-api"
version = "0.1.0"
description = "An Open edX plugin providing a public API for course outline (modules summary) for product and marketing pages"
authors = [
  {name = "MIT Office of Digital Learning"}
]
license = "BSD-3-Clause"
readme = "README.rst"
requires-python = ">=3.11"
dependencies = [
  "Django>=4.0",
  "djangorestframework>=3.14.0",
  "edx-django-utils>4.0.0",
  "edx-drf-extensions>=10.0.0",
  "edx-opaque-keys",
]

[project.entry-points."lms.djangoapp"]
ol_openedx_course_outline_api = "ol_openedx_course_outline_api.app:CourseOutlineAPIConfig"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ol_openedx_course_outline_api"]
include = [
  "ol_openedx_course_outline_api/**/*.py",
]

[tool.hatch.build.targets.sdist]
include = [
  "ol_openedx_course_outline_api/**/*",
  "README.rst",
  "pyproject.toml",
]
